feat(dock-manager): build sample for hiding panes #48
