+++ browser/public/index.html
@@ -330,6 +330,7 @@
        <label id="nav-dock-manager" class="nav-component">Dock Manager</label> 
        <div id="nav-dock-manager-list" class="nav-list" state="collapsed"> 
            <a class="nav-link" href="#" data-nav="/samples/layouts/dock-manager-embedding-frames">Embedding Frames</a> 
+           <a class="nav-link" href="#" data-nav="/samples/layouts/dock-manager-hiding-panes">Hide Panes</a> 
            <a class="nav-link" href="#" data-nav="/samples/layouts/dock-manager-overview">Overview</a> 
            <a class="nav-link" href="#" data-nav="/samples/layouts/dock-manager-updating-panes">Updating Panes</a> 
        </div>
